break up failing arkane tests, fix one, skip the other
broke the two arkane input tests into separate files, one which covers using arkane in a functional programming style and one which uses it through input files.
The latter I was able to get working by adding global declarations, the former is broken in a similar way to the explorerTest, which is also skipped
